refactor: modifier classes into multiple classes
